add test for blockage Campbells Creek 50
initially did not work.  not because of change 1 --> 50, but because TravelTimeCollector (now renamed WithinDayTravelTime) did not include network change events from before qsim start, AND the qsim start fo the 50 example is _later_ than for the _1_ example.  A bit of bad luck in this particular situation, but better to catch this now in general.